Providing_decide_nudges: Chanegd code to work on new functions, skipping out the nudge_select() function
nudge_list: Changed function to combine it with nudge_select() so that this one function does both jobs. This involved:
tweaking function to be able to retun a raw number of nudges which involved adding an ifelse() statement.
Added a  way of specifying the weights column. The data frame that's output now has a column with whatever this term is.

nudge_select: changed the name of the weighting_column argument to make it more logical. This function has now been made redundant because of the changes to nudge_list()